refactor(render): split render_native_tests.cpp into focused test files
Split the 1571-line test file into 8 files: render_test_helpers (shared
assertions), test_style_tokens, test_style_sequencer, test_theme_blend,
test_formatting, test_qt_hooks, test_c_api_errors, and slimmed runner.
No behavioral changes — pure file reorganization.
